# docs/api-spec.yaml
openapi: 3.0.3
info:
  title: Kaizen API
  version: 0.1.0
  description: >
    REST API for Kaizen. Single source of truth for endpoints and schemas.
    Payloads use snake_case. All protected routes require a Bearer JWT.
    MVP scope = auth, items, streaks, sync, rule-based redistribute.
    AI endpoints are Phase 1 (premium) and are tagged accordingly.
servers:
  - url: http://localhost:3000
    description: Local dev
  - url: https://api.kaizen.com
    description: Production (placeholder)
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/subscription/dev-upgrade:
    post:
      tags: [Subscription]
      summary: DEV ONLY — set the current user's subscription tier (non-production)
      description: >
        Flips subscription_tier without real payment so premium features (AI)
        can be tested before payments (RevenueCat, Phase 1) are integrated.
        Returns 404 when NODE_ENV=production.
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                tier: { type: string, enum: [free, premium], default: premium }
      responses:
        '200':
          description: Updated user
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { description: Not available (production) }
  /api/v1/subscription/status:
    get:
      tags: [Subscription]
      summary: Get the current user's premium entitlement status (ADR-041)
      description: >
        Single source of truth for "am I premium?" — works regardless of
        the billing channel (Apple IAP / Google Play / RuStore / Stripe /
        YooKassa / dev-upgrade). Uses server-side resolveEntitlement:
        isPremium = subscriptionTier="premium" OR (premiumUntil != null AND
        premiumUntil > now). The app calls this on startup and after any
        purchase attempt.
      responses:
        '200':
          description: Entitlement status
          content:
            application/json:
              schema:
                type: object
                required: [is_premium]
                properties:
                  is_premium:
                    type: boolean
                    description: True if the user has an active premium subscription.
                  premium_until:
                    type: string
                    format: date-time
                    nullable: true
                    description: >
                      Expiry of the timed subscription (null for legacy
                      lifetime premium or if no subscription is active).
                  source:
                    type: string
                    nullable: true
                    enum: [apple, google, rustore, stripe, yookassa, dev, null]
                    description: Billing channel that granted premium.
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/billing/webhook/apple:
    post:
      tags: [Billing]
      summary: Apple App Store webhook stub (ADR-041)
      description: >
        Stub webhook — sets premiumUntil / premiumSource for the user.
        Real implementation requires App Store Server Notifications V2 JWS
        signature verification (TODO when Apple Developer account is available).
        If BILLING_WEBHOOK_SECRET env is set, requires x-webhook-secret header.
      parameters:
        - in: header
          name: x-webhook-secret
          schema: { type: string }
          description: Required if BILLING_WEBHOOK_SECRET env is set.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/BillingWebhookRequest' }
      responses:
        '200':
          description: Entitlement updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BillingWebhookResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/billing/webhook/google:
    post:
      tags: [Billing]
      summary: Google Play Billing webhook stub (ADR-041)
      description: >
        Stub webhook. Real implementation requires JWT verification with
        Google service account key (TODO).
        If BILLING_WEBHOOK_SECRET env is set, requires x-webhook-secret header.
      parameters:
        - in: header
          name: x-webhook-secret
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/BillingWebhookRequest' }
      responses:
        '200':
          description: Entitlement updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BillingWebhookResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/billing/webhook/rustore:
    post:
      tags: [Billing]
      summary: RuStore Billing webhook stub (ADR-041)
      description: >
        Stub webhook. Real implementation requires HMAC verification with
        BILLING_RUSTORE_SECRET (TODO).
        If BILLING_WEBHOOK_SECRET env is set, requires x-webhook-secret header.
      parameters:
        - in: header
          name: x-webhook-secret
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/BillingWebhookRequest' }
      responses:
        '200':
          description: Entitlement updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BillingWebhookResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/billing/webhook/stripe:
    post:
      tags: [Billing]
      summary: Stripe webhook stub (ADR-041)
      description: >
        Stub webhook. Real implementation requires Stripe-Signature header
        verification via stripe.webhooks.constructEvent (TODO when Stripe
        account is available).
        If BILLING_WEBHOOK_SECRET env is set, requires x-webhook-secret header.
      parameters:
        - in: header
          name: x-webhook-secret
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/BillingWebhookRequest' }
      responses:
        '200':
          description: Entitlement updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BillingWebhookResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/billing/webhook/yookassa:
    post:
      tags: [Billing]
      summary: YooKassa webhook stub (ADR-041)
      description: >
        Stub webhook. Real implementation requires IP allowlist check +
        SHA-1 signature verification (TODO when YooKassa account is available).
        If BILLING_WEBHOOK_SECRET env is set, requires x-webhook-secret header.
      parameters:
        - in: header
          name: x-webhook-secret
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/BillingWebhookRequest' }
      responses:
        '200':
          description: Entitlement updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BillingWebhookResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    BadRequest:
      description: Validation error
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Unauthorized:
      description: Missing or invalid JWT
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    NotFound:
      description: Not found (also returned for items owned by another user)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string, example: Email already exists }
    Tone:
      type: string
      enum: [gentle, harsh]
...
    User:
      type: object
      description: passwordHash is never returned
      properties:
        id: { type: string, format: uuid }
        email: { type: string, format: email, nullable: true }
        phone: { type: string, nullable: true, description: 'Russian phone, E.164: +7XXXXXXXXXX' }
        name: { type: string }
        subscription_tier: { type: string, enum: [free, premium] }
        is_premium:
          type: boolean
          description: >
            Resolved entitlement (ADR-041): true if subscriptionTier="premium"
            (legacy) OR (premiumUntil is set AND premiumUntil > now).
        premium_until:
          type: string
          format: date-time
          nullable: true
          description: Expiry of the timed subscription; null for legacy lifetime or no subscription.
        premium_source:
          type: string
          nullable: true
          enum: [apple, google, rustore, stripe, yookassa, dev, null]
          description: Billing channel that granted premium.
        theme: { type: string, enum: [focus, calm, black, white, contrast] }
        tone_preference: { $ref: '#/components/schemas/Tone' }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    BillingWebhookRequest:
      type: object
      required: [user_id, expires_at]
      description: >
        Stub payload accepted by all billing webhook endpoints (ADR-041).
        In real integrations the provider sends its own envelope; this stub
        provides a uniform dev/test surface. All fields are snake_case.
      properties:
        user_id:
          type: string
          description: The Kaizen user ID to update (uuid).
        product_id:
          type: string
          description: Optional SKU / product identifier for logging.
        expires_at:
          type: string
          format: date-time
          description: ISO 8601 expiry timestamp for the subscription.
    BillingWebhookResponse:
      type: object
      required: [ok]
      properties:
        ok: { type: boolean, example: true }
...